Create README.md
Here I have tried to create a sample application as an example to showcase integration of Struts2, Spring, JQuery, Ajax and Bootstrap.  
This is a simple web application for inventory management of a store. The user would a "Store Manager" or "admin"( not customers) who would be able to login and view product list in a datagrid table and perform basic CRUD operations like adding, updating and removing products from list. The login info for user is stored in a CSV file within the application and used for user's login authentication. 
There is no database used for storing the items in the products, so application uses sysnchronized lists to make sure multiple users are able to view the same list of products regardless of their own individual session. The list will be maintained as long the application remains active on the server and has not been deleted by any user.
A user can also upload image and view them in the data table. The application has some basic authentication functionality on the client-side using JQuery validation. 
Hope this helps. Happy coding. Cheers! :)